+++ publisher/src/test/scala/org/podval/tools/publish/markup/FootnoteSpec.scala
@@ -2,7 +2,6 @@ package org.podval.tools.publish.markup
 
 import org.podval.tools.publish.site.{PageError, PageErrorReporter}
 import org.podval.xml.{HtmlXmlWriterConfig, Xml, XmlElement, XmlParser}
-import Xml.given
 import org.scalatest.funsuite.AnyFunSuite
 import zio.blocks.chunk.Chunk