+++ publisher/src/test/scala/org/podval/tools/publish/page/PagedListSpec.scala
@@ -1,7 +1,6 @@
 package org.podval.tools.publish.page
 
 import org.podval.xml.{HtmlXmlWriterConfig, Xml}
-import Xml.given
 import org.scalatest.funsuite.AnyFunSuite
 
 final class PagedListSpec extends AnyFunSuite: